requirements.txt
Flask
google-generativeai
python-dotenv
markdown